mv CommandArgs into Message as MessageParams
